+++ python/GafferUI/__init__.py
@@ -184,6 +184,7 @@ from . import _Pointer
 from .SplineWidget import SplineWidget
 from .Bookmarks import Bookmarks
 from . import WidgetAlgo
+from .CodeWidget import CodeWidget
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.